Build C files with -fno-builtin
The built-in memcpy and memset functions appear to
assume sizeof(int) == 4, and sizeof(size_t) == 2.
This seems wrong in general for size_t, but is
wrong for int when using -mshort as well. To avoid
this issue, build with -fno-builtin to prevent gcc
from inserting built-in-compatible function calls
that conflict with the types of the non-builtin
locally defined equivalents.