Fix and size-optimize nvmskunk.s
Too many changes to list. The FLUSH and Getbytes
routines were hooked into a unit test harness and
massaged until they worked properly, then the
entire file was optimized for size.

The main issue with the previous routines was that
the Skunkboard flash chip's address lines do not
match up with their corresponding lines on the
Jaguar. This is fine as long as all the addresses
in a given 64k sector correspond to other
addresses in that same sector (meaning flash
word-address lines A0-A14 correspond to Jaguar
byte-address lines EA1-EA15), but that is not the
case for A12 and A14, which are mapped to EA20 and
EA19 respectively. This is still fine as long as
individual sectors aren't erased and written
independently of other sectors, which is how the
Skunkboard is used when you're just erasing 2MB or
4MB chunks and flashing ROMs over that space.
The memory track firmware on the other hand needs
to be able to address sectors correctly to allow
applications to read and write small files. To fix
this, whenever the Skunkboard memory track flash
driver accesses an addresses, the address is first
"fixed" by swizzling the correct Jaguar address
bits into the lower 64k range of the address.

See skunk_addr_mapping.txt for details of the
address line mappings derived from the Skunkboard
PCB.