Move NVRAM content to 2MB boundary in 2nd bank
Because of the address swizzling required to
correctly access sectors, the NVRAM data seems
to be overwriting the boot ROM code, causing
subsequent boots to fail. To avoid this, move
it to the 2nd bank. Users will still be able
to utilize their first bank for other games,
but they'll need to re-flash the memory track
ROM afterwards to access their save games again.