Give the ROM a header the Skunkboard can recognize
By necessity, the memory track BIOS ROM will be
auto-booted when the user activates the CD boot
mode by holding down 'C' while booting the
Skunkboard in a Jaguar CD unit. To ensure the
Skunkboard BIOS doesn't boot some random other
ROM, give this one a unique signature the
Skunkboard BIOS can check for, similar to the
existing auto-boot code in bank 2, but which the
BIOS will only look for when booting in CD mode.