Copy Skunk BIOS to bank 2 if needed
If the 2nd bank of the Skunkboard appears to be
blank where a copy of the BIOS should be, copy
it over. This allows the board to boot properly
when left in bank 2 and reset. Since the memory
track BIOS leaves the board in bank 2, this is
quite useful.

This still doesn't properly handle the case where
someone installed the memory track ROM over a 6MB
ROM, because in that case the BIOS area in bank 2
may not be blank and will not contain BIOS data.
That will take a little more effort to fix, so
defer it for now.